Ditto for set (allow adding properties to the prototype)

+++ sbbs3/getmsg.cpp
@@ -2,13 +2,13 @@
 
 /* Synchronet message retrieval functions */
 
-/* $Id: getmsg.cpp,v 1.38 2010/03/06 00:13:04 rswindell Exp $ */
+/* $Id: getmsg.cpp,v 1.37 2009/10/17 23:19:37 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
  * @format.use-tabs true	(see http://www.synchro.net/ptsc_hdr.html)		*
  *																			*
- * Copyright 2010 Rob Swindell - http://www.synchro.net/copyright.html		*
+ * Copyright 2009 Rob Swindell - http://www.synchro.net/copyright.html		*
  *																			*
  * This program is free software; you can redistribute it and/or			*
  * modify it under the terms of the GNU General Public License				*
@@ -255,7 +255,7 @@ ulong sbbs_t::getmsgnum(uint subnum, time_t t)
 		return(0); 
 	}
 
-	total=(long)filelength(fileno(smb.sid_fp))/sizeof(idxrec_t);
+	total=filelength(fileno(smb.sid_fp))/sizeof(idxrec_t);
 
 	if(!total) {		   /* Empty base */
 		smb_close(&smb);
@@ -416,7 +416,7 @@ ulong sbbs_t::getlastmsg(uint subnum, uint32_t *ptr, time_t *t)
 		errormsg(WHERE,ERR_READ,smb.file,i,smb.last_error);
 		return(0); 
 	}
-	total=(long)filelength(fileno(smb.sid_fp))/sizeof(idxrec_t);
+	total=filelength(fileno(smb.sid_fp))/sizeof(idxrec_t);
 	smb_unlocksmbhdr(&smb);
 	smb_close(&smb);
 	if(ptr)
